move functions from base to Inheritance
